Separate oracle and noir_js encoding
