+++ packages/noir-ethereum-api/src/noir/oracles/encode.test.ts
